Fix acknowledgment system in pace.c and make visual changes
All detectors and SensorSignalReader() use the acknowledgment system
at all times (not only when the signal is activated). Otherwise, it is
possible for a signal to change it's values this way 1->0->1 and the
detectors may not notice the change if they don't read the signal while
it's value is 0. That way oldValues[] isn't updated and the signal is
lost.

Other changes:
- toggle_signal() calculates the return value beforehand when using
bitfields.
- Use a temporary variable for signalArray[target] in the inside loop
of detectors. Otherwise, the target element could change (without the
acknowledgment system). This is also very important for
BitfieldChangeDetector() because the acknowledgment system doesn't
"protect" the other 31 elements in the target bitfield.